Moved code to 'jdk.jsplitpkgscan' to fix #38
In additon also the module name was changed to the same name as the
package and fixed build bradge link.
