--- build.xml
@@ -1,26 +1,0 @@
-<project name="jsplitpkgscan" default="dist">
-  <property name="src" location="src"/>
-  <property name="build" location="build/classes/main"/>
-  <property name="dist" location="build/libs"/>
-
-  <target name="init">
-    <tstamp/>
-    <mkdir dir="${build}"/>
-  </target>
-
-  <target name="compile" depends="init" description="compile the source">
-    <javac srcdir="${src}/main/java" destdir="${build}"/>
-  </target>
-
-  <target name="dist" depends="compile" description="generate the distribution">
-    <mkdir dir="${dist}"/>
-    <jar jarfile="${dist}/jsplitpkgscan-${DSTAMP}.jar" basedir="${build}">
-      <fileset dir="${src}/main/resources"/>
-    </jar>
-  </target>
-
-  <target name="clean" description="clean up">
-    <delete dir="${build}"/>
-    <delete dir="${dist}"/>
-  </target>
-</project>
