+++ src/main/java/module-info.java
@@ -1,7 +1,7 @@
-module org.adoptopenjdk.jsplitpkgscan {
+module jdk.jsplitpkgscan {
     requires java.compiler;
     provides java.util.spi.ToolProvider
-            with jdk.Main.JsplitpgkscanToolProvider;
+            with jdk.jsplitpkgscan.Main.JsplitpgkscanToolProvider;
     provides javax.tools.Tool
-            with jdk.Main.JsplitpgkscanTool;
+            with jdk.jsplitpkgscan.Main.JsplitpgkscanTool;
 }
