+++ src/main/resources/jdk/jsplitpkgscan/resources/jsplitpkgscan.properties
